GRMHD_cpp: Fixing the PDE. It works now (with 80% certainty of my side).
The (major) problem was that the initial data (AlfenWave) were for 3D but the PDE was compiled in 2D. The PDE works now always in 3D (as the Fortran GRMHD PDE did). This really improves everything.

Long term flat space stability tests and a small convergence series are still todo, but now at least we probably have a working PDE.