Adding new application "PotentialHydro".
This was an old experiment from Jul 28, 2017. Actually mimicing a Poisson equation (Newtonian physics) whith a hyperbolic PDE (i.e. going from 3 to 4 dimensions and decreasing the propagation speed of the field from infinity to some final but large value) is something people actually do. Anyway, this PDE is abandoned so far.