BuildScripts: Reparing compile.sh to go with the specfile-centric world view.
compile.sh can no more be standalone and depends now on the exa tools. On the other hand, the massive variables get stripped down to a few defining ones (in the end, only the absolute specfile path). This should be changed even more.

I bet the Out Of Tree build system is completely broken, too.