Plotters: Adding new ADERDG2LegendreCSV plotter.
This class was there already before, as a debugging class used by Vasco. I rewrote it in order to get a bit compatibility with the CarpetASCII file format (not yet archieved) and to get some output in textual form.

Code compiles but is untested, next thing is to test it, integrate slicing operations and offer cartesian interpolation.

I would especially like to offer the option for approximate slicing so one can see computational vertices.