GRMHD: Major upgrade.
We now have a powerful constants system integrated for ID. The Shocktube ID generator uses it already to dynamically setup shocktubes from input parameters (in contrast to hard coded parameters we had before).

To perform this setp, the ID API was rewritten, furthermore constants where added on the DG and FV solver.

The whole code is untested. And the BC still miss an integration into the new API.